requirements.txt
streamlit
torch
transformers
faiss-cpu
Pillow
requests
tqdm
